refactor(governance): moves proposal under /[genre]
