+++ apps/hub/src/app/governance/[genre]/proposal/[proposalId]/components/queue-button.tsx
@@ -2,12 +2,14 @@ import { GOVERNANCE_ABI, useBeraJs } from "@bera/berajs";
 import { governanceTimelockAddress } from "@bera/config";
 import { ActionButton, useTxn } from "@bera/shared-ui";
 import { Button } from "@bera/ui/button";
+import { useGovernance } from "~/app/governance/[genre]/components/governance-provider";
 
 export const QueueButton = ({ proposalId }: { proposalId: string }) => {
   const { write, ModalPortal } = useTxn({
     message: "Queuing proposal",
   });
 
+  const { governorAddress } = useGovernance();
   return (
     <>
       {ModalPortal}
@@ -15,7 +17,7 @@ export const QueueButton = ({ proposalId }: { proposalId: string }) => {
         <Button
           onClick={() =>
             write({
-              address: governanceTimelockAddress,
+              address: governorAddress,
               abi: GOVERNANCE_ABI,
               functionName: "queue",
               params: [BigInt(proposalId)],
